+++ Heimdall/Heimdall.Application/Contracts/IHeimdallConfigValidator.cs
@@ -1,0 +1,8 @@
+﻿using Heimdall.Application.Configuration;
+
+namespace Heimdall.Application.Contracts;
+
+public interface IHeimdallConfigValidator
+{
+    void ValidateAndThrow(HeimdallConfig config);
+}