+++ Heimdall/Heimdall.Application/Contracts/IRunSummaryService.cs
@@ -1,0 +1,8 @@
+﻿using Heimdall.Domain.Results;
+
+namespace Heimdall.Application.Contracts;
+
+public interface IRunSummaryService
+{
+    string RenderText(RunSummary runSummary);
+}